Winui (#5)
* feat: add frontend that follows the style of WinUI 3 and offers a design/layout similar to the Windows App.

Features: favorites, desktops list with wallpaper for each desktop, apps list sorted by folders, choosing between terminal servers when an app is available on multiple, favorites (with importing and exporting), simple view (similar to the old frontend), automatic dark mode, folder flattening, sign out button, splash screen while loading, stale while revalidate.

* feat: add webmanifest to enable PWA with window controls overlay support and an install prompt

* fix: use logo svg from /app/lib/assets within app

* fix: omit critical files and other non-app resources from the service worker cache

* feat: add option to hide square behind icons

* feat: add TextBox component

* feat: add sort, view, and search options on each page of sevices and/or apps

* refactor: combine RemoteAppCard and GenericResourceCard

* fix: correctly specify columns for resource grid list mode

* fix: open rdp file on resource card click

* fix: do not bubble click events from content dialogs

Content dialogs appear in a popover, so their context seems like ti should be at teh top level. Therefore, their click events should not propogate to their paret components since the popovers are visually detatched.

* fix: do not show hover or active effect on resource card when clicking or tapping the menu

* fix: actually vertically center active menu flyout item idicator

* feat: show menu divider between the two secitons of the sort dropdown

* feat: show resource card menu when card is focused by the keyboard

* feat: open resource card menu on right click

* feat: account for webfeeds that do not follow the requirement to group remoteapps with the same connection parameters other than terminal server/address

* fix: use proper search bar label

* fix: apply search query and sort isntructions to apps list when flatten folders is enabled

* feat: respect prefers-reduce-motion

* feat: add offline support

* feat: persist non-data localstorage when switching users

* fix: use correct timestamp for resources

Resource timestamps should indicate when the entire resource has updated, not just the rdp file. That means all related image files must also be included.

* fix: use unique identifiers for resource id and alias attribute values

* fix: prevent duplicate resources

If the RDP file contents are exactly the same, they should not appear multiple times in the webfeed. This breaks some RDP clients because it does not follow the specification.

* fix: derive terminal servers from rdp file `fullAddress` instead of using the hosting server name

* fix: use correct publisher timestamp based on last modified timestamp of all resources

* fix: do not specify `FileAssociationIcons`  if the icon does not exist

Before, it always included an IconRaw element with a URL to the ICO file for the file extension, even if the icon for the file extension did not exist. IconRaw should not exist when the file extension icon does not exist. Always including this icon meant that the file extensions received the default icon even when the RemoteApp may have its own icon.

* fix: do not include FileAssociationIcons with schema versions < 2

* fix: omit PrimaryHandler attribute from FileExtension on schema versions < 2

* fix: follow 2.2.2.2.6 - specify multiple folders within a single resource instead of sending duplicate resources

* fix: include ShowByDefault attribute or Resource elements for schema versions 2.1

* fix: include all valid icon elements

* fix: use FQDN for publisher id instead of server name

* fix: ignore duplicate resources in the same folder

Resources that repeat multiple times the same folder (e.g., file1.rdp and file2.rdp with same file contents) should be ignored after the first resource is included. For schema version 1, this means using the folder path instead of the file path when generating the GUID. For schema version 2, this means checking whether the folder has already been included in the list of folders.

* adjust code now that RAWeb is MS-TSWP compliant

* feat: add login page (with GUI)

* feat: add logoff page

* feat: support desktop icons based on custom wallpaper

For desktops, provide a png file of the wallpaper. RAWeb will put the wallpaper inside a frame of a PC, similar to the design of the This PC icon on Windows 11.

* feat: use modern default.ico with windows 11 style

* fix: wait for cache to be cleared before completing logoff

* fix: show correct resource name label (device for desktops, application for remoteapps)

* fix: do not filter our desktops on simple mode resources page

* fix: do not trigger resource card right click events from dialog surfaces

* fix: resize images that are too big for ico output format

The logic we use to convert PNG to ICO does not work when the image is larger than 256x256. We now resize the image to fit within the 256x256 size while maintaining the aspect ratio.

* feat: show loading indicator in titlebar when service worker is fetching data in the background

* feat: combine resources that are the same other than their terminal server when `?combineTerminalServers=1` is included in the URL

* fix: actually respect the `maxSize` parameter in `GetResizedDimensionsFromMaxSize`

* fix: force a new webapp login when existing credentials become invalid

* feat: add settings toggle to control mergeTerminalServers option for webfeed in the web app

* feat: show server name in titlebar

* feat: show machine name on login page

* feat: hide domain for username in titlebar

* feat: use terminal server aliases for apps and desktops in the web app

* feat: add terminal servers filter to apps list

* feat: search apps and devices automatically

* fix: do not omit get-image.aspx from service worker

* feat: open help in a popup instead of a new tab when left clicking

* feat: use default wallpaper instead of default icon for desktops without an icon

* feat: configure IIS to not set long cache headers for files that will be cached by the service worker

* feat: add IIS appConfig options to force web app options

* fix: do not import non-existent RemoteAppCard component

This removes the error in the browser.

* feat: improve keyboard navigation

Every element should now respond to enter key and space key. All interactive elements will show an outline when focused by the keyboard. Opening a menu flyout will focus the first item in the menu flyout. Arrow keys are now used to navigate menu flyouts. The context menu keyboard button now triggers the right click menu on resource cards instead of needing to tab to the menu icon.

* refactor: move IIS-related variables to `window` from `app.appContext.config.globalProperties`

* fix: provide progressring size as a number

* refactor: move most of main app root doc to its own control

* refactor: move frontend code to new folder

* feat: add option to install raweb frontend with vite or simple copy

* fix: do not merge desktops even when mergeTerminalServers=1

the mergeTerminalServers functionality was only meant for remoteapps

* update screenshots

* chore: fix webfeed.aspx after merge

* feat!: replace old frontend with new frontend